Remove .NET Standard 2.0 target and set .NET Core 3.1 as the minimum target
