+++ Directory.Build.props
@@ -48,6 +48,7 @@
     <SystemDiagnosticsPerformanceCounterVersion>5.0.0</SystemDiagnosticsPerformanceCounterVersion>
     <SystemNetNameResolutionVersion>4.3.0</SystemNetNameResolutionVersion>
     <MicrosoftDotNetPlatformAbstractionsVersion>3.1.6</MicrosoftDotNetPlatformAbstractionsVersion>
+    <SystemSecurityCryptographyCngVersion>5.0.0</SystemSecurityCryptographyCngVersion>
 
     <!-- Microsoft packages -->
     <MicrosoftBuildVersion>16.8.0</MicrosoftBuildVersion>
@@ -92,7 +93,7 @@
     <NewRelicAgentApiVersion>8.0.0.0</NewRelicAgentApiVersion>
     <NewtonsoftJsonVersion>12.0.3</NewtonsoftJsonVersion>
     <ZooKeeperNetExVersion>3.4.12.4</ZooKeeperNetExVersion>
-    <StackExchangeRedis>2.0.601</StackExchangeRedis>
+    <StackExchangeRedisVersion>2.0.601</StackExchangeRedisVersion>
     <KubernetesClientVersion>4.0.5</KubernetesClientVersion>
 
     <!-- Test related packages -->
