+++ Orleans.sln
@@ -128,8 +128,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Reminders.DynamoDB"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Streaming.SQS", "src\AWS\Orleans.Streaming.SQS\Orleans.Streaming.SQS.csproj", "{DA0FC102-756F-4C1F-8018-FEC5A64F06F6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetCore.Tests", "test\NetCore.Tests\NetCore.Tests.csproj", "{1F28D680-7B22-4EA7-BA92-163D4FD76F5E}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Azure", "Azure", "{4C5D66BF-EE1C-4DD8-8551-D1B7F3768A34}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "AWS", "AWS", "{DA8E126B-BCDB-4E8F-BFB9-2DBFD41F8F70}"
@@ -442,10 +440,6 @@ Global
 		{DA0FC102-756F-4C1F-8018-FEC5A64F06F6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DA0FC102-756F-4C1F-8018-FEC5A64F06F6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DA0FC102-756F-4C1F-8018-FEC5A64F06F6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1F28D680-7B22-4EA7-BA92-163D4FD76F5E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1F28D680-7B22-4EA7-BA92-163D4FD76F5E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1F28D680-7B22-4EA7-BA92-163D4FD76F5E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1F28D680-7B22-4EA7-BA92-163D4FD76F5E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{CE21A15A-E376-4B81-BDA6-3AB610CFE03A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CE21A15A-E376-4B81-BDA6-3AB610CFE03A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CE21A15A-E376-4B81-BDA6-3AB610CFE03A}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -612,7 +606,6 @@ Global
 		{4D298A28-FF46-49EC-B8A4-A758389E5DF3} = {DA8E126B-BCDB-4E8F-BFB9-2DBFD41F8F70}
 		{99BAAA1B-1C5A-461D-B1E8-1937D9738BA1} = {DA8E126B-BCDB-4E8F-BFB9-2DBFD41F8F70}
 		{DA0FC102-756F-4C1F-8018-FEC5A64F06F6} = {DA8E126B-BCDB-4E8F-BFB9-2DBFD41F8F70}
-		{1F28D680-7B22-4EA7-BA92-163D4FD76F5E} = {A6573187-FD0D-4DF7-91D1-03E07E470C0A}
 		{4C5D66BF-EE1C-4DD8-8551-D1B7F3768A34} = {FE2E08C6-9C3B-4AEE-AE07-CCA387580D7A}
 		{DA8E126B-BCDB-4E8F-BFB9-2DBFD41F8F70} = {FE2E08C6-9C3B-4AEE-AE07-CCA387580D7A}
 		{CE21A15A-E376-4B81-BDA6-3AB610CFE03A} = {EB2EDE59-5021-42EE-A97A-D59939B39C66}
