+++ src/Orleans.Streaming/Orleans.Streaming.csproj
@@ -5,5 +5,6 @@
     <Description>Streaming library for Microsoft Orleans used both on the client and server.</Description>
     <TargetFrameworks>$(MultiTargetFrameworks)</TargetFrameworks>
     <OrleansBuildTimeCodeGen>true</OrleansBuildTimeCodeGen>
+    <ProduceReferenceAssembly>false</ProduceReferenceAssembly>
   </PropertyGroup>
 </Project>
