+++ test/Misc/TestInterfaces/TestInterfaces.csproj
@@ -2,6 +2,6 @@
   <PropertyGroup>
     <RootNamespace>UnitTests.Interfaces</RootNamespace>
     <AssemblyName>TestInterfaces</AssemblyName>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 </Project>
