--- test/NetCore.Tests/NetCore.Tests.xunit.runner.json
@@ -1,8 +1,0 @@
-{
-  "appDomain": "denied",
-  "diagnosticMessages": true,
-  "parallelizeAssembly": false,
-  "parallelizeTestCollections": false,
-  "methodDisplay": "classAndMethod",
-  "shadowCopy": false
-}
