--- test/NetCore.Tests/Properties/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-using Xunit;
-
-// Disable XUnit concurrency limit
-[assembly: CollectionBehavior(MaxParallelThreads = -1)]
